moved all executor function into own dir called executor, also added headers to all files in this dir
