+++ src/parser/dispatcher.c
@@ -2,7 +2,7 @@
 #include <commands/commands.h>
 #include <commands/echo_utils.h>
 #include <parser/parser.h>
-#include <output/executor_utils.h>
+#include <executor/executor_utils.h>
 #include <stdio.h>
 #include <libft.h>
 #include <ctype.h>
